+++ to edit/drawable/rect.h
@@ -13,27 +13,27 @@ class Rect
 {
 	
 protected:
-	int width;
-	int height;
-	int left;
-	int top;
+	float width;
+	float height;
+	float left;
+	float top;
 	
 	
 public:
 
-	void set( int x, int y, int w, int h );
+	void set( float x, float y, float w, float h );
 	
 	// accessors
-	const int& getX() const;
-	const int& getY() const;
-    const int& getWidth() const;
-    const int& getHeight() const;
+	const float& getX() const;
+	const float& getY() const;
+    const float& getWidth() const;
+    const float& getHeight() const;
 	
-    const int& getLeft() const;
-    const int getRight() const;
-    const int& getTop() const;
-    const int getBot() const;
+    const float& getLeft() const;
+    const float getRight() const;
+    const float& getTop() const;
+    const float getBot() const;
 	
-	bool checkCollision( int x, int y, int w = 0, int h = 0 );
+	bool checkCollision( float x, float y, float w = 0, float h = 0 );
 	bool checkRectCollision( Rect rect );
 };
